requirements.txt
Django==2.0.2
requests==2.18.4
Pillow==4.2.1
boto==2.48.0
boto3==1.4.7
opencv-python==3.2.0.8
gcs-oauth2-boto-plugin==1.14
google-api-core==0.1.1
google-api-python-client==1.6.4
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-storage==1.6.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
whitenoise==3.3.1
imgurpython==1.1.7
django-ratelimit==1.1.0
celery==4.1.0
redis==2.10.6
celery[redis]